+++ common/src/main/java/com/github/cao/awa/conium/bedrock/entity/BedrockEntity.kt
@@ -2,21 +2,38 @@ package com.github.cao.awa.conium.bedrock.entity
 
 import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApi
 import com.github.cao.awa.conium.annotation.bedrock.BedrockScriptApiFacade
+import com.github.cao.awa.conium.annotation.script.javascript.ScriptReadonly
+import com.github.cao.awa.conium.bedrock.block.state.bedrock
+import com.github.cao.awa.conium.bedrock.raycast.hit.BlockRaycastHit
 import com.github.cao.awa.conium.bedrock.script.BedrockScriptAnonymousObjectMap
+import com.github.cao.awa.conium.bedrock.world.dimension.BedrockDimension
+import com.github.cao.awa.conium.bedrock.world.dimension.bedrockDimension
+import com.github.cao.awa.conium.kotlin.extent.innate.int
+import com.github.cao.awa.conium.kotlin.extent.innate.orGetAuto
 import com.github.cao.awa.conium.kotlin.extent.world.executeCommand
+import com.github.cao.awa.conium.raycast.ConiumRaycast
 import com.github.cao.awa.sinuatum.util.collection.CollectionFactor
 import net.minecraft.entity.Entity
-import net.minecraft.server.MinecraftServer
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.server.world.ServerWorld
+import net.minecraft.util.hit.HitResult
+import net.minecraft.util.math.BlockPos
+import net.minecraft.util.math.Vec3d
+import net.minecraft.world.RaycastContext
+import net.minecraft.world.World
 import org.apache.logging.log4j.LogManager
+import org.apache.logging.log4j.Logger
 
 @BedrockScriptApi
 @BedrockScriptApiFacade("Entity")
 open class BedrockEntity(private val delegate: Entity) {
     companion object {
-        private val LOGGER = LogManager.getLogger("BedrockEntity")
+        private val LOGGER: Logger = LogManager.getLogger("BedrockEntity")
     }
+
+    @ScriptReadonly
+    @BedrockScriptApiFacade("Entity", "#dimension")
+    val dimension: BedrockDimension = this.delegate.world.bedrockDimension
 
     @BedrockScriptApiFacade("Entity", "teleport")
     fun teleport(location: BedrockScriptAnonymousObjectMap, teleportOption: BedrockScriptAnonymousObjectMap) {
@@ -42,9 +59,47 @@ open class BedrockEntity(private val delegate: Entity) {
                     this.delegate,
                     command
                 )
-            } else{
+            } else {
                 LOGGER.warn("Unable to execute command '$command' as entity that are not player")
             }
+        }
+    }
+
+    @BedrockScriptApiFacade("Entity", "runCommandAsync")
+    fun runCommandAsync(command: String) = runCommand(command)
+
+    @BedrockScriptApiFacade("Entity", "getBlockFromViewDirection")
+    fun getBlockFromViewDirection(options: BedrockScriptAnonymousObjectMap? = BedrockScriptAnonymousObjectMap.EMPTY): BlockRaycastHit? {
+        val maxDistance: Double = options.orGetAuto(20.0) { it["maxDistance"] }
+        val includeLiquidBlocks: Boolean = options.orGetAuto(false) { it["includeLiquidBlocks"] }
+        val includePassableBlocks: Boolean = options.orGetAuto(false) { it["includePassableBlocks"] }
+
+        ConiumRaycast.raycast(
+            this.delegate,
+            maxDistance,
+            0F,
+            includeLiquidBlocks,
+            includePassableBlocks
+        ).let { hitResult: HitResult ->
+            val hitType: HitResult.Type = hitResult.type
+            val pos: Vec3d = hitResult.pos
+            val world: World = this.delegate.world
+
+            if (hitType == HitResult.Type.MISS || hitType == HitResult.Type.ENTITY) {
+                return null
+            }
+
+            val blockPos = BlockPos(
+                pos.x.int,
+                pos.y.int,
+                pos.z.int,
+            )
+
+            return BlockRaycastHit(
+                world.getBlockState(blockPos).bedrock,
+                this.delegate.facing,
+                blockPos
+            )
         }
     }
 
@@ -53,4 +108,4 @@ open class BedrockEntity(private val delegate: Entity) {
     }
 }
 
-fun Entity.toBedrock(): BedrockEntity = BedrockEntity(this)
+fun Entity.bedrock(): BedrockEntity = BedrockEntity(this)