+++ common/src/main/java/com/github/cao/awa/conium/bedrock/index/BedrockImportIndexes.kt
@@ -21,8 +21,8 @@ import com.github.cao.awa.conium.bedrock.entity.player.screen.BedrockOnScreenDis
 // Event.
 import com.github.cao.awa.conium.bedrock.event.context.BedrockEventContext
 import com.github.cao.awa.conium.bedrock.event.context.item.use.BedrockItemUseOnEventContext
-import com.github.cao.awa.conium.bedrock.event.before.item.use.BedrockItemUseOnBeforeEvent
-import com.github.cao.awa.conium.bedrock.event.after.item.use.BedrockItemUseOnAfterEvent
+import com.github.cao.awa.conium.bedrock.event.before.item.use.on.BedrockItemUseOnBeforeEvent
+import com.github.cao.awa.conium.bedrock.event.after.item.use.on.BedrockItemUseOnAfterEvent
 
 // Item.
 import com.github.cao.awa.conium.bedrock.item.stack.BedrockItemStack