+++ common/src/main/java/com/github/cao/awa/conium/kotlin/extent/innate/ConiumInnate.kt
@@ -13,6 +13,20 @@ inline fun <reified T> T.changeIfIs(value: T, creator: (T) -> T): T {
     return this
 }
 
+inline fun <reified T, reified R> T?.orGet(value: R, op: (T) -> R): R {
+    if (this == null) {
+        return value
+    }
+    return op(this)
+}
+
+inline fun <reified T, reified R> T?.orGetAuto(value: R, op: (T) -> Any?): R {
+    if (this == null) {
+        return value
+    }
+    return op(this) as R
+}
+
 inline val Number.int: Int get() = toInt()
 inline val Number.long: Long get() = toLong()
 inline val Number.float: Float get() = toFloat()